requirements.txt
asgiref==3.8.1
Django==5.1.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
PyJWT==2.10.0
sqlparse==0.5.2
tzdata==2024.2
